+++ events/management/commands/initialize_edx_common_events.py
@@ -1,6 +1,6 @@
 from django.core.management.base import BaseCommand
 
-from events.enums import RggInternalEventTypes
+from events.enums import EdxCommonEventTypes
 from events.models import EventType, EventConfiguration
 
 
@@ -8,15 +8,16 @@ class Command(BaseCommand):
     """
     Management command for creating Event Types and Event Configurations.
 
-    Create or update RGG internal Event Types and Event Configurations
+    Create Event Types and Event Configurations according to common edX events.
     """
 
-    help = 'Creates EventType and EventConfiguration entries for RggInternalEventTypes enum.'
+    help = 'Creates EventType and EventConfiguration entries for EdxCommonEventTypes enum.'
 
     def handle(self, *args, **options):
-        for event in RggInternalEventTypes:
+        for event in EdxCommonEventTypes:
             event_name = event.value
             title = event.title
+            award = event.award
 
             event_type, created_type = EventType.objects.get_or_create(name=event_name)
             if created_type:
@@ -24,14 +25,14 @@ class Command(BaseCommand):
             else:
                 self.stdout.write(f'EventType already exists: {event_name}')
 
-            event_config, created_config = EventConfiguration.objects.update_or_create(
+            event_config, created_config = EventConfiguration.objects.get_or_create(
                 event_type=event_type,
                 defaults={
                     'title': title,
-                    'award': 0,
+                    'award': award,
                 }
             )
             if created_config:
                 self.stdout.write(self.style.SUCCESS(f'Created EventConfiguration for: {title}'))
             else:
-                self.stdout.write(self.style.SUCCESS(f'Updated EventConfiguration (award reset to 0) for: {title}'))
+                self.stdout.write(f'EventConfiguration already exists: {title}')