+++ events/management/tests/test_commands.py
@@ -1,7 +1,7 @@
 import pytest
 from django.core.management import call_command
 
-from events.enums import RggInternalEventTypes
+from events.enums import EdxCommonEventTypes, RggInternalEventTypes
 from events.models import EventType, EventConfiguration
 
 
@@ -73,3 +73,76 @@ def test_initialize_rgg_internal_events_reset_award_to_zero(
 
     assert EventType.objects.count() == len(RggInternalEventTypes.get_all())
     assert EventConfiguration.objects.count() == len(RggInternalEventTypes.get_all())
+
+
+@pytest.mark.no_rgg_events
+@pytest.mark.django_db
+def test_initialize_edx_common_events_non_existing():
+    call_command('initialize_edx_common_events')
+
+    for enum_item in EdxCommonEventTypes:
+        event_type = EventType.objects.get(name=enum_item.value)
+        event_configuration = EventConfiguration.objects.get(event_type=event_type)
+
+        assert event_type.name == enum_item.value
+        assert event_configuration.title == enum_item.title
+        assert event_configuration.award == enum_item.award
+
+    assert EventType.objects.count() == len(EdxCommonEventTypes.get_all())
+    assert EventConfiguration.objects.count() == len(EdxCommonEventTypes.get_all())
+
+
+@pytest.mark.no_rgg_events
+@pytest.mark.django_db
+def test_initialize_edx_common_events_does_not_duplicate_existing_event_type_and_configuration(
+    event_type_factory, event_configuration_factory
+):
+    for enum_item in EdxCommonEventTypes:
+        event_type = event_type_factory(name=enum_item.value)
+        event_configuration_factory(event_type=event_type, title=enum_item.title)
+
+    assert EventType.objects.count() == len(EdxCommonEventTypes.get_all())
+    assert EventConfiguration.objects.count() == len(EdxCommonEventTypes.get_all())
+
+    call_command('initialize_edx_common_events')
+
+    for enum_item in EdxCommonEventTypes:
+        event_type = EventType.objects.get(name=enum_item.value)
+        event_configuration = EventConfiguration.objects.get(event_type=event_type)
+
+        assert event_type.name == enum_item.value
+        assert event_configuration.title == enum_item.title
+
+    assert EventType.objects.count() == len(EdxCommonEventTypes.get_all())
+    assert EventConfiguration.objects.count() == len(EdxCommonEventTypes.get_all())
+
+
+@pytest.mark.no_rgg_events
+@pytest.mark.django_db
+def test_initialize_edx_common_events_not_reset_existent_award_and_title(
+    event_type_factory, event_configuration_factory
+):
+    for counter, enum_item in enumerate(EdxCommonEventTypes):
+        event_type = event_type_factory(name=enum_item.value)
+        event_configuration = event_configuration_factory(
+            event_type=event_type, award=100, title=f'test_title_{counter}'
+        )
+
+        assert event_configuration.award == 100
+        assert event_configuration.title == f'test_title_{counter}'
+
+    assert EventType.objects.count() == len(EdxCommonEventTypes.get_all())
+    assert EventConfiguration.objects.count() == len(EdxCommonEventTypes.get_all())
+
+    call_command('initialize_edx_common_events')
+
+    for counter, enum_item in enumerate(EdxCommonEventTypes):
+        event_type = EventType.objects.get(name=enum_item.value)
+        event_configuration = EventConfiguration.objects.get(event_type=event_type)
+
+        assert event_type.name == enum_item.value
+        assert event_configuration.title == f'test_title_{counter}'
+        assert event_configuration.award == 100
+
+    assert EventType.objects.count() == len(EdxCommonEventTypes.get_all())
+    assert EventConfiguration.objects.count() == len(EdxCommonEventTypes.get_all())